core/standards/scripts/scheduler2/location_hash/312.html

<!doctype html>
<title>Location hash - set hash in external script with defer - delayed 1 second</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});
</script>
<script>
t.step(function() {
    location.hash = '#foo';
});
</script>
<script src="312.php?delay=1" id="test_script" defer></script>
<script>
setTimeout(before_completely_loaded, 100);
function before_completely_loaded() {
    t.step(function() {
        assert_equals( location.hash, '#foo' );
        assert_equals( history.length, histlength );
    });
    t.done();
}
</script>
<div id="log"></div>
